Add pdbinfo source and executable
pdbinfo is a simple tool that extracts the timestamp, GUID, and age from
a PDB file. The GUID and age can then be compared against the GUID and
age in a binary to make sure that they match.

Typical usage of pdbinfo looks like this:

C:\>pdbinfo pdbinfo.pdb
Timestamp            GUID                        age
5760F503, {B4B698F7-5C8A-4AC2-B8E6-8475E81FB693}, 1

Typical usage of dumpbin to get the matching information looks like this:

C:\>"%VS140COMNTOOLS%..\..\VC\vcvarsall.bat" amd64
C:\>dumpbin pdbinfo.exe /headers | find /i "rsds"
  5760F503 cv  44 000037CC  1DCC  Format: RSDS, {B4B698F7-5C8A-4AC2-B8E6-8475E81FB693}, 1, pdbinfo.pdb

Note that the timestamps are not guaranteed to match, but the GUID and age are.