Source code and binaries for Alt+Tab fixers
Windows 7's Explorer has a bug that causes it to mess up the display of
the Alt+Tab list of programs, sometimes triggered by always-on-top,
visible, zero size windows. These programs hack around this problem by
making the windows invisible.

The issue and the fix are described in greater detail here:

https://randomascii.wordpress.com/2013/05/28/alttab-bug-worsened-by-ie-10-but-fix-found/