Better handle .error files when trimming symbols
trim_symbols.py cleans up old .pdb files in symbol servers so that they
don't grow forever. It looks for files that we have multiple versions of
and deletes the old ones.

It tries to be careful not to delete files that aren't clearly part of a
symbol server directory, and it contained logic to handle .error files,
which are left behind by failed downloads. However it only allowed for
there being a single .error file in an otherwise empty directory. There
can be multiple .error files, in the same directory as a successful
download. Apparently the symbol server download logic never cleans up
after itself.

So, the trim_symbols.py script now deletes any .error files that it
finds and then rescans to see if the directory looks okay. This let it
clean up an extra 4.7 GB (!) representing 2% of the total space on my
laptop's SSD. There were 1.7 GB of .error files (13 of them) and 3.0
GB of .pdb files (10 of them) which were locked in an undeletable state
by the .error files and the script's old logic.