Add script to display random wallpaper
Windows 10 took away the handy "random good wallpaper" and apparently it
took me five years to get around to recreating it. This script makes use
of my existing database of photos, with star ratings listed. It filters
that database (actually just a text file) down to just the five-star
photos, skips the videos, and then selects one of those great photos as
the desktop wallpaper. It also records which photo was selected because
it's often nice to know where a photo is from.

Because this is dependent on the photo database it is not much use to
anyone else, but maybe some day I'll share that code as well.