Updates to RandomWallpaper and new photo script
This updates RandomWallpaper.py to add a --showall option to tell the
script to select from all of the photos instead of just the five-star
photos. Maybe that should be the default? This also improves some
comments, moves the timing so that it prints the timing regardless of
how execution leaves main().

CopyPhotoSubset.py uses a different photo database to find photos with
particular tags and copy these to another directory.

With these two scripts I copied all photos featuring a particular person
to their laptop and set it up to cycle through them - a cheap photo
frame for the vain.